Reorganized the data loaders
These files have grown too large. I have split them up to make them a bit more manageable.

ECOM-5099
