+++ course_discovery/apps/course_metadata/data_loaders/tests/__init__.py
@@ -1,0 +1,1 @@
+JSON = 'application/json'
